+++ README.md
@@ -44,4 +44,4 @@
 
 ### 요구사항 3 - Controller 인터페이스 추가하기
 
-- [ ] 컨트롤러 인터페이스를 추가하고 각 분기에 있는 로직마다 AbstractController를 상속한 구현체로 만들어보자.
+- [x] 컨트롤러 인터페이스를 추가하고 각 분기에 있는 로직마다 AbstractController를 상속한 구현체로 만들어보자.
