+++ tomcat/src/main/java/camp/nextstep/http/HttpStatus.java
@@ -6,6 +6,7 @@ public enum HttpStatus {
   FOUND(302, "Found"),
   UNAUTHORIZED(401, "Unauthorized"),
   NOT_FOUND(404, "Not Found"),
+  METHOD_NOT_ALLOWED(405, "Method Not Allowed"),
   INTERNAL_SERVER_ERROR(500, "Internal Server Error"),
   ;
   private final int statusCode;
